Second Commit

+++ src/main.rs
@@ -1,6 +1,6 @@
 use clap::Parser;
 use chrono::*;
-use std::{io::{Write, Read}, fs::File};
+use std::{io::{Write, Read}, path::Path};
 
 #[derive(Parser)]
 #[clap(version, about)]
@@ -15,39 +15,41 @@ struct Cli {
     add: bool,
 }
 
-fn make_file() -> File {
-    let mut app = match std::fs::File::create("rtrc") {
+fn make_file(){
+    let mut app = match std::fs::File::create("~/.config/rtrc") {
         Ok(file) => file,
         Err(_) => panic!("Failed to make file"),
     };
     app.write_all(b"What ever is below CREATOR MESSAGE will be added to the created file.\n\n").expect("Write failed.");
     app.write_all(b"[Creator Message]\n").expect("Write failed.");
-    return app;
+    println!("Please configure the \"~/.config/.rtrc\"");
 }
 
 fn make_file_with_text(args: Cli) {
-    let mut text_file = match std::fs::File::open("rtrc") {
-        Ok(file) => file,
-        Err(_) => make_file()
-    };
-    let mut file = std::fs::File::create(args.path).expect("Error making file");
-    let mut contents = String::new();
-    text_file.read_to_string(&mut contents).expect("Failed to read config file.");
-    let mut state = false;
-    for line in contents.lines() {
-        if line == "[Language Comment]" { break; }
-        if line.ends_with("Date: ") {
-            file.write_all(format!("{}", line).as_bytes()).expect("Failed to write");
-            file.write_all(format!("{}\n", Local::now().date()).as_bytes()).expect("Failed to write");
-            continue;
+    if Path::new("~/.config/rtrc").exists() {
+        let mut text_file = std::fs::File::open("~/.config/rtrc").unwrap();
+        let mut file = std::fs::File::create(args.path).expect("Error making file");
+        let mut contents = String::new();
+        text_file.read_to_string(&mut contents).expect("Failed to read config file.");
+        let mut state = false;
+        for line in contents.lines() {
+            if line == "[Language Comment]" { break; }
+            if line.ends_with("Date: ") {
+                file.write_all(format!("{}", line).as_bytes()).expect("Failed to write");
+                file.write_all(format!("{}\n", Local::now().date()).as_bytes()).expect("Failed to write");
+                continue;
+            }
+            if state {
+                file.write_all(format!("{}\n", line).as_bytes()).expect("Failed to write");
+                continue;
+            }
+            if line == "[Creator Message]" {
+                state = true;
+            }
         }
-        if state {
-            file.write_all(format!("{}\n", line).as_bytes()).expect("Failed to write");
-            continue;
-        }
-        if line == "[Creator Message]" {
-            state = true;
-        }
+    } else {
+        make_file();
+        println!("Please configure the .config/.rtrc");
     }
 }
